ai_agent_src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pydantic==2.5.0
pydantic-settings==2.1.0
redis[asyncio]==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.13.1
SQLAlchemy==2.0.36
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0